feat(6690): extract residual root service api tests
