+++ crates/kamn-node/src/main_tests/service_api_endpoint_tests/residual_root_contract_tests.rs
@@ -1,0 +1,12 @@
+use super::*;
+
+#[path = "residual_root_contract_tests/env_lock_contract_tests.rs"]
+mod env_lock_contract_tests;
+#[path = "residual_root_contract_tests/serde_payload_contract_tests.rs"]
+mod serde_payload_contract_tests;
+#[path = "residual_root_contract_tests/error_envelope_contract_tests.rs"]
+mod error_envelope_contract_tests;
+#[path = "residual_root_contract_tests/payload_parse_contract_tests.rs"]
+mod payload_parse_contract_tests;
+#[path = "residual_root_contract_tests/missing_resource_contract_tests.rs"]
+mod missing_resource_contract_tests;
